% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/random_puzzle.R
\name{apply_random_positioning}
\alias{apply_random_positioning}
\title{Apply positioning to random shape puzzle}
\usage{
apply_random_positioning(piece_result, offset)
}
\arguments{
\item{piece_result}{Result from generate_random_pieces_internal()}

\item{offset}{Separation amount (0 = compact, >0 = separated)}
}
\value{
Positioned piece result
}
\description{
Separates pieces based on offset parameter.
}
\keyword{internal}
