% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/piece_positioning.R
\name{apply_rect_positioning}
\alias{apply_rect_positioning}
\title{Apply rectangular piece positioning with offset}
\usage{
apply_rect_positioning(piece_result, offset)
}
\arguments{
\item{piece_result}{Output from generate_pieces_internal() for rectangular}

\item{offset}{Separation distance in mm}
}
\value{
Positioned result
}
\description{
Apply rectangular piece positioning with offset
}
\keyword{internal}
