% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rcpp_wrappers.R
\name{bezier_batch}
\alias{bezier_batch}
\title{Compute cubic Bezier curve points}
\usage{
bezier_batch(p0, cp1, cp2, p1, n_points = 20)
}
\arguments{
\item{p0}{Start point as c(x, y)}

\item{cp1}{First control point as c(x, y)}

\item{cp2}{Second control point as c(x, y)}

\item{p1}{End point as c(x, y)}

\item{n_points}{Number of points to generate (default 20)}
}
\value{
Data frame with x and y columns
}
\description{
Uses C++ implementation when available, falling back to R otherwise.
Computes points along the curve defined by P0, CP1, CP2, P1.
}
\keyword{internal}
