% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{bezier_batch_cpp}
\alias{bezier_batch_cpp}
\title{Compute cubic Bezier curve points in batch}
\usage{
bezier_batch_cpp(p0, cp1, cp2, p1, n_points)
}
\arguments{
\item{p0}{Start point as c(x, y)}

\item{cp1}{First control point as c(x, y)}

\item{cp2}{Second control point as c(x, y)}

\item{p1}{End point as c(x, y)}

\item{n_points}{Number of points to generate}
}
\value{
NumericMatrix with x and y columns
}
\description{
Vectorized implementation of cubic Bezier interpolation.
Computes points along the curve defined by P0, CP1, CP2, P1.
}
\keyword{internal}
