% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bezier_utils.R
\name{bezier_to_points}
\alias{bezier_to_points}
\title{Convert cubic Bezier curve to line segments}
\usage{
bezier_to_points(p0, cp1, cp2, p1, n_points = 20)
}
\arguments{
\item{p0}{Start point as c(x, y)}

\item{cp1}{First control point as c(x, y)}

\item{cp2}{Second control point as c(x, y)}

\item{p1}{End point as c(x, y)}

\item{n_points}{Number of points to generate (default 20)}
}
\value{
Data frame with x and y columns
}
\description{
Uses De Casteljau's algorithm to approximate a cubic Bezier curve
as a series of line segments suitable for polygon rendering.
}
\examples{
# Simple S-curve
pts <- bezier_to_points(c(0, 0), c(0, 1), c(1, 0), c(1, 1))
plot(pts$x, pts$y, type = "l")
}
