% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/puzzle_core_clean.R
\name{blend_colors}
\alias{blend_colors}
\title{Blend two hex colors by averaging RGB values}
\usage{
blend_colors(color1, color2)
}
\arguments{
\item{color1}{First hex color}

\item{color2}{Second hex color}
}
\value{
Blended hex color string
}
\description{
Blend two hex colors by averaging RGB values
}
