% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/concentric_edge_generation.R
\name{build_concentric_piece_path}
\alias{build_concentric_piece_path}
\title{Build SVG path for a concentric piece}
\usage{
build_concentric_piece_path(piece_id, edge_data)
}
\arguments{
\item{piece_id}{Piece ID}

\item{edge_data}{Edge map data from generate_concentric_edge_map}
}
\value{
SVG path string
}
\description{
Build SVG path for a concentric piece
}
