% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hexagonal_individual_pieces.R
\name{calculate_hexagonal_neighbors}
\alias{calculate_hexagonal_neighbors}
\title{Calculate hexagonal neighbors for a piece at (q, r)}
\usage{
calculate_hexagonal_neighbors(q, r)
}
\arguments{
\item{q}{Axial coordinate q}

\item{r}{Axial coordinate r}
}
\value{
List of neighbor coordinates
}
\description{
Calculate hexagonal neighbors for a piece at (q, r)
}
