% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hexagonal_separation.R
\name{calculate_hexagonal_piece_bounds}
\alias{calculate_hexagonal_piece_bounds}
\title{Calculate bounds for hexagonal pieces}
\usage{
calculate_hexagonal_piece_bounds(pieces)
}
\arguments{
\item{pieces}{List of hexagonal pieces}
}
\value{
List of bounding box data for each piece
}
\description{
Calculate bounds for hexagonal pieces
}
