% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hexagonal_individual_pieces.R
\name{calculate_hexagonal_piece_positions}
\alias{calculate_hexagonal_piece_positions}
\title{Calculate hexagonal piece positions using proper axial coordinates}
\usage{
calculate_hexagonal_piece_positions(rings)
}
\arguments{
\item{rings}{Number of rings in the puzzle}
}
\value{
List of piece position data with proper hexagonal coordinates
}
\description{
Determines where each piece is located in the hexagonal grid using
axial coordinates (q, r) and calculates their center points.
}
\details{
For rings=3, we should have:
- Ring 0: 1 piece (center)
- Ring 1: 6 pieces  
- Ring 2: 12 pieces
- Total: 19 pieces
}
