% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/puzzle_separation.R
\name{calculate_optimal_offset}
\alias{calculate_optimal_offset}
\title{Calculate optimal offset for laser cutting}
\usage{
calculate_optimal_offset(
  piece_width,
  piece_height,
  kerf = 0.2,
  min_separation = 2
)
}
\arguments{
\item{piece_width}{Average width of pieces}

\item{piece_height}{Average height of pieces}

\item{kerf}{Laser kerf width (material removed by cutting)}

\item{min_separation}{Minimum separation between pieces}
}
\value{
Recommended offset value
}
\description{
Determines appropriate separation based on piece size and kerf.
}
