% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/random_puzzle.R
\name{calculate_random_bounds}
\alias{calculate_random_bounds}
\title{Calculate bounds for random shape pieces}
\usage{
calculate_random_bounds(pieces)
}
\arguments{
\item{pieces}{List of piece objects}
}
\value{
List with min_x, max_x, min_y, max_y
}
\description{
Calculate bounds for random shape pieces
}
\keyword{internal}
