% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/piece_positioning.R
\name{calculate_separated_canvas}
\alias{calculate_separated_canvas}
\title{Calculate canvas size for separated layout}
\usage{
calculate_separated_canvas(positioned)
}
\arguments{
\item{positioned}{Output from apply_piece_positioning()}
}
\value{
c(width, height)
}
\description{
Calculate canvas size for separated layout
}
\keyword{internal}
