% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/piece_positioning.R
\name{calculate_tessellation_bounds}
\alias{calculate_tessellation_bounds}
\title{Calculate bounds for tessellation pieces}
\usage{
calculate_tessellation_bounds(pieces)
}
\arguments{
\item{pieces}{List of piece objects}
}
\value{
List with min_x, max_x, min_y, max_y
}
\description{
Calculate bounds for tessellation pieces
}
\keyword{internal}
