% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/voronoi_puzzle.R
\name{calculate_voronoi_bounds}
\alias{calculate_voronoi_bounds}
\title{Calculate bounds for Voronoi pieces}
\usage{
calculate_voronoi_bounds(pieces)
}
\arguments{
\item{pieces}{List of piece objects}
}
\value{
List with min_x, max_x, min_y, max_y
}
\description{
Calculate bounds for Voronoi pieces
}
\keyword{internal}
