% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/noise_fills.R
\name{check_ambient_available}
\alias{check_ambient_available}
\title{Check if ambient package is available}
\usage{
check_ambient_available()
}
\value{
TRUE if all required packages are available
}
\description{
Check if ambient package is available
}
\keyword{internal}
