% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/launch_app.R
\name{check_app_dependencies}
\alias{check_app_dependencies}
\title{Check Shiny App Dependencies}
\usage{
check_app_dependencies()
}
\value{
Logical indicating if all dependencies are met
}
\description{
Checks if all required packages for the Shiny app are installed
}
\examples{
check_app_dependencies()

}
