% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/piece_positioning.R
\name{compute_bbox_overlap}
\alias{compute_bbox_overlap}
\title{Compute overlap between two bounding boxes}
\usage{
compute_bbox_overlap(bbox1, bbox2, margin = 0)
}
\arguments{
\item{bbox1}{First bounding box}

\item{bbox2}{Second bounding box}

\item{margin}{Required gap between boxes}
}
\value{
List with overlaps (boolean), overlap_x, overlap_y
}
\description{
Compute overlap between two bounding boxes
}
\keyword{internal}
