% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adjacency_api.R
\name{compute_fused_edges}
\alias{compute_fused_edges}
\title{Compute fused edges from fusion groups}
\usage{
compute_fused_edges(fusion_groups, puzzle_result)
}
\arguments{
\item{fusion_groups}{List of integer vectors (piece IDs to fuse)}

\item{puzzle_result}{Output from generate_puzzle()}
}
\value{
List with:
  \describe{
    \item{fused_edges}{Character vector of edge keys to fuse}
    \item{edge_to_group}{Named list mapping edge keys to fusion group index}
    \item{piece_to_group}{Named list mapping piece IDs to fusion group index}
  }
}
\description{
Converts fusion groups to a set of edge keys that should be treated as
"fused" (internal edges within meta-pieces).
}
