% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/piece_positioning.R
\name{compute_group_bbox}
\alias{compute_group_bbox}
\title{Compute bounding box for a group of pieces}
\usage{
compute_group_bbox(pieces)
}
\arguments{
\item{pieces}{List of piece objects}
}
\value{
List with min_x, max_x, min_y, max_y, center_x, center_y
}
\description{
Compute bounding box for a group of pieces
}
\keyword{internal}
