% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/piece_positioning.R
\name{compute_repulsion_vector}
\alias{compute_repulsion_vector}
\title{Compute repulsion vector between two overlapping boxes}
\usage{
compute_repulsion_vector(bbox1, bbox2, overlap, step_size = 1)
}
\arguments{
\item{bbox1}{First bounding box}

\item{bbox2}{Second bounding box}

\item{overlap}{Overlap info from compute_bbox_overlap}

\item{step_size}{Multiplier for repulsion distance}
}
\value{
c(dx, dy) repulsion vector (from bbox1 to bbox2)
}
\description{
Compute repulsion vector between two overlapping boxes
}
\keyword{internal}
