% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/image_processing.R
\name{convert_svg_to_png}
\alias{convert_svg_to_png}
\title{Convert SVG to PNG using available tools}
\usage{
convert_svg_to_png(svg_content, output_file, width_px = 2000, height_px = 2000)
}
\arguments{
\item{svg_content}{SVG content as string}

\item{output_file}{Output PNG filename}

\item{width_px}{Width in pixels}

\item{height_px}{Height in pixels}
}
\value{
TRUE if successful, FALSE otherwise
}
\description{
Convert SVG to PNG using available tools
}
