% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_individual_pieces_proper.R
\name{create_combined_view}
\alias{create_combined_view}
\title{Create combined view of all pieces}
\usage{
create_combined_view(pieces, width, height, output_dir)
}
\description{
Create combined view of all pieces
}
\keyword{internal}
