% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hexagonal_individual_pieces.R
\name{create_fallback_hexagonal_path}
\alias{create_fallback_hexagonal_path}
\title{Create fallback hexagonal path when segments unavailable}
\usage{
create_fallback_hexagonal_path(piece_pos, diameter, rings)
}
\arguments{
\item{piece_pos}{Piece position data}

\item{diameter}{Puzzle diameter}

\item{rings}{Number of rings}
}
\value{
Simple hexagonal path
}
\description{
Create fallback hexagonal path when segments unavailable
}
