% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/config_utils.R
\name{detect_puzzle_type}
\alias{detect_puzzle_type}
\title{Detect puzzle type from pieces structure}
\usage{
detect_puzzle_type(pieces)
}
\arguments{
\item{pieces}{List of piece objects}
}
\value{
Character string: "rectangular", "hexagonal", "concentric", or "other"
}
\description{
Examines piece attributes to determine the puzzle type.
}
\keyword{internal}
