% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/individual_pieces_final.R
\name{.generate_curved_edge_forward}
\alias{.generate_curved_edge_forward}
\title{Internal function to generate curved edge using tab parameters}
\usage{
.generate_curved_edge_forward(tab_params)
}
\description{
Internal function to generate curved edge using tab parameters
}
\keyword{internal}
