% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/noise_fills.R
\name{encode_noise_png}
\alias{encode_noise_png}
\title{Encode color matrix as base64 PNG}
\usage{
encode_noise_png(color_matrix, width, height)
}
\arguments{
\item{color_matrix}{3D array (height x width x 3) of RGB values [0,1]}

\item{width}{Image width}

\item{height}{Image height}
}
\value{
Base64-encoded PNG data URI string
}
\description{
Uses the png package to write the color matrix to a PNG file
and encodes it as a base64 data URI.
}
\keyword{internal}
