% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hexagonal_individual_pieces.R
\name{extract_coordinates_from_path}
\alias{extract_coordinates_from_path}
\title{Extract coordinates from SVG path string}
\usage{
extract_coordinates_from_path(path)
}
\arguments{
\item{path}{SVG path string}
}
\value{
Numeric vector of coordinates
}
\description{
Extract coordinates from SVG path string
}
