% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/random_puzzle.R
\name{extract_random_adjacency}
\alias{extract_random_adjacency}
\title{Extract adjacency from triangulation}
\usage{
extract_random_adjacency(tri, vertices, n_corner)
}
\arguments{
\item{tri}{RCDT triangulation result}

\item{vertices}{All vertices matrix}

\item{n_corner}{Number of boundary vertices}
}
\value{
Data frame with cell_a, cell_b, v1_x, v1_y, v2_x, v2_y columns
}
\description{
Builds adjacency data from RCDT triangulation output.
}
\keyword{internal}
