% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/individual_pieces.R
\name{extract_tab_data}
\alias{extract_tab_data}
\title{Extract tab/blank data from puzzle generation}
\usage{
extract_tab_data(seed, xn, yn, tabsize = 20, jitter = 4)
}
\arguments{
\item{seed}{Random seed for puzzle generation}

\item{xn}{Number of columns}

\item{yn}{Number of rows}

\item{tabsize}{Tab size percentage}

\item{jitter}{Jitter percentage}
}
\value{
List containing tab/blank data for all dividers
}
\description{
This function pre-generates all tab/blank information needed to create
individual piece boundaries consistently.
}
