% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hexagonal_individual_pieces.R
\name{find_border_segment_for_edge_piece}
\alias{find_border_segment_for_edge_piece}
\title{Find border segment for edge piece}
\usage{
find_border_segment_for_edge_piece(piece_pos, path_segments, rings)
}
\arguments{
\item{piece_pos}{Edge piece position data}

\item{path_segments}{All parsed path segments}

\item{rings}{Number of rings}
}
\value{
Border segment for the edge piece
}
\description{
Find border segment for edge piece
}
