% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hexagonal_piece_extraction.R
\name{find_connecting_segments}
\alias{find_connecting_segments}
\title{Find segments that connect to a given endpoint}
\usage{
find_connecting_segments(point, segments, tolerance = 0.1)
}
\arguments{
\item{point}{Target point c(x, y)}

\item{segments}{List of segments to search}

\item{tolerance}{Matching tolerance for coordinates}
}
\value{
Indices of matching segments
}
\description{
Find segments that connect to a given endpoint
}
