% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/renderer_edge_splitting.R
\name{find_path_vertices}
\alias{find_path_vertices}
\title{Find vertices (corners) in a path by detecting angle changes}
\usage{
find_path_vertices(endpoints, n_edges)
}
\arguments{
\item{endpoints}{List of c(x, y) endpoint coordinates}

\item{n_edges}{Expected number of edges (4 for trapezoid, 6 for hexagon)}
}
\value{
List of vertex indices in the endpoints list
}
\description{
Find vertices (corners) in a path by detecting angle changes
}
\keyword{internal}
