% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hexagonal_individual_pieces.R
\name{find_segment_between_hex_pieces}
\alias{find_segment_between_hex_pieces}
\title{Find the path segment connecting two hexagonal pieces}
\usage{
find_segment_between_hex_pieces(piece1, piece2, path_segments, rings)
}
\arguments{
\item{piece1}{First piece position data}

\item{piece2}{Second piece position data (neighbor)}

\item{path_segments}{All parsed path segments}

\item{rings}{Number of rings}
}
\value{
Segment connecting the pieces, or NULL if none found
}
\description{
Find the path segment connecting two hexagonal pieces
}
