% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hexagonal_piece_extraction.R
\name{generate_separated_hex_real}
\alias{generate_separated_hex_real}
\title{Generate separated hexagonal puzzle with REAL pieces (work in progress)}
\usage{
generate_separated_hex_real(
  rings = 3,
  seed = NULL,
  diameter = 240,
  offset = 10,
  tabsize = 27,
  jitter = 5,
  do_warp = FALSE,
  do_trunc = FALSE,
  colors = NULL,
  stroke_width = 1.5
)
}
\arguments{
\item{rings}{Number of rings}

\item{seed}{Random seed}

\item{diameter}{Puzzle diameter}

\item{offset}{Separation distance}

\item{tabsize}{Tab size percentage}

\item{jitter}{Jitter percentage}

\item{do_warp}{Apply circular warping}

\item{do_trunc}{Truncate edge pieces}

\item{colors}{Piece colors}

\item{stroke_width}{Line width}
}
\value{
SVG content with separated real pieces
}
\description{
Generate separated hexagonal puzzle with REAL pieces (work in progress)
}
