% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hexagonal_separation.R
\name{generate_separated_hexagonal_svg}
\alias{generate_separated_hexagonal_svg}
\title{Generate separated hexagonal puzzle SVG}
\usage{
generate_separated_hexagonal_svg(
  rings = 3,
  seed = NULL,
  diameter = 240,
  offset = 10,
  arrangement = "hexagonal",
  tabsize = 27,
  jitter = 5,
  do_warp = FALSE,
  do_trunc = FALSE,
  colors = NULL,
  stroke_width = 1,
  background = "none"
)
}
\arguments{
\item{rings}{Number of rings}

\item{seed}{Random seed}

\item{diameter}{Puzzle diameter}

\item{offset}{Separation distance between pieces}

\item{arrangement}{"hexagonal" or "rectangular" packing}

\item{tabsize}{Tab size percentage}

\item{jitter}{Jitter percentage}

\item{do_warp}{Apply circular warping}

\item{do_trunc}{Truncate edge pieces}

\item{colors}{Piece colors}

\item{stroke_width}{Line width}

\item{background}{Background color or "none"}
}
\value{
SVG content as string
}
\description{
Creates an SVG with hexagonal puzzle pieces separated for laser cutting.
}
