% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/concentric_geometry.R
\name{get_all_concentric_vertices}
\alias{get_all_concentric_vertices}
\title{Get all vertices for all pieces in a concentric puzzle}
\usage{
get_all_concentric_vertices(rings, diameter, center_shape = "hexagon")
}
\arguments{
\item{rings}{Number of rings}

\item{diameter}{Puzzle diameter}

\item{center_shape}{"hexagon" or "circle"}
}
\value{
List of piece vertex data
}
\description{
Get all vertices for all pieces in a concentric puzzle
}
