% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/piles_notation.R
\name{get_boundary_pieces}
\alias{get_boundary_pieces}
\title{Get boundary pieces (internal)}
\usage{
get_boundary_pieces(puzzle_result = NULL)
}
\arguments{
\item{puzzle_result}{Puzzle result}
}
\value{
Integer vector of boundary piece IDs
}
\description{
Get boundary pieces (internal)
}
\keyword{internal}
