% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/piles_notation.R
\name{get_col_pieces}
\alias{get_col_pieces}
\title{Get pieces in a specific column (rectangular puzzles) (internal)}
\usage{
get_col_pieces(col_num, puzzle_result = NULL)
}
\arguments{
\item{col_num}{Column number (1-indexed from left)}

\item{puzzle_result}{Puzzle result}
}
\value{
Integer vector of piece IDs
}
\description{
Get pieces in a specific column (rectangular puzzles) (internal)
}
\keyword{internal}
