% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adjacency_api.R
\name{get_concentric_neighbors}
\alias{get_concentric_neighbors}
\title{Get all neighbors for a concentric puzzle piece}
\usage{
get_concentric_neighbors(piece_id, rings)
}
\arguments{
\item{piece_id}{Piece ID (1-based)}

\item{rings}{Number of rings in puzzle}
}
\value{
Data frame with direction, neighbor_id, is_boundary
}
\description{
Returns neighbor information for all edges of a concentric piece.
Center piece (ring 0) has 6 edges, trapezoid pieces have 4 edges.
}
