% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/concentric_geometry.R
\name{get_concentric_piece_count}
\alias{get_concentric_piece_count}
\title{Calculate the number of pieces for a concentric ring puzzle}
\usage{
get_concentric_piece_count(rings)
}
\arguments{
\item{rings}{Number of rings (including center)}
}
\value{
Total number of pieces
}
\description{
Calculate the number of pieces for a concentric ring puzzle
}
