% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/concentric_geometry.R
\name{get_concentric_piece_height}
\alias{get_concentric_piece_height}
\title{Calculate piece height (radial extent) for concentric mode}
\usage{
get_concentric_piece_height(diameter, rings)
}
\arguments{
\item{diameter}{Puzzle diameter}

\item{rings}{Number of rings}
}
\value{
Piece height in same units as diameter
}
\description{
Calculate piece height (radial extent) for concentric mode
}
