% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adjacency_api.R
\name{get_fusion_group_pieces}
\alias{get_fusion_group_pieces}
\title{Get all pieces in a fusion group}
\usage{
get_fusion_group_pieces(group_idx, fusion_groups)
}
\arguments{
\item{group_idx}{Fusion group index}

\item{fusion_groups}{Original fusion groups list}
}
\value{
Vector of piece IDs in the group
}
\description{
Get all pieces in a fusion group
}
