% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adjacency_api.R
\name{get_group_internal_edges}
\alias{get_group_internal_edges}
\title{Get internal edges of a fusion group}
\usage{
get_group_internal_edges(group_idx, fused_edge_data)
}
\arguments{
\item{group_idx}{Fusion group index}

\item{fused_edge_data}{Output from compute_fused_edges()}
}
\value{
Character vector of edge keys
}
\description{
Returns all edges that are internal to a fusion group (edges between
pieces within the same group).
}
