% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adjacency_api.R
\name{get_hex_complementary_side}
\alias{get_hex_complementary_side}
\title{Get complementary side for hexagonal puzzle}
\usage{
get_hex_complementary_side(side)
}
\arguments{
\item{side}{Side number (0-5) as character or integer}
}
\value{
Opposite side number as character
}
\description{
Returns the opposite side number. For hexagons, the opposite of side n
is (n + 3) mod 6.
}
