% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hexagonal_adjacency_cache.R
\name{get_hex_neighbors_fast}
\alias{get_hex_neighbors_fast}
\title{Get all neighbors for a piece using cached matrix}
\usage{
get_hex_neighbors_fast(piece_id, rings)
}
\arguments{
\item{piece_id}{Piece ID (1 to num_pieces)}

\item{rings}{Number of rings in puzzle}
}
\value{
Data frame with direction, neighbor_id, is_boundary
}
\description{
Returns pre-allocated data frame with all 6 neighbors.
Much faster than iterative rbind approach.
}
