% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adjacency_api.R
\name{get_hex_topo_neighbor_matrix}
\alias{get_hex_topo_neighbor_matrix}
\title{Get cached neighbor matrix (legacy interface)}
\usage{
get_hex_topo_neighbor_matrix(rings)
}
\arguments{
\item{rings}{Number of rings in puzzle}
}
\value{
Matrix where matrix[piece_id, side+1] = neighbor_id (or NA)
}
\description{
Get cached neighbor matrix (legacy interface)
}
\keyword{internal}
