% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/concentric_edge_generation.R
\name{get_inner_neighbor}
\alias{get_inner_neighbor}
\title{Get inner ring piece ID that contains the midpoint (legacy function)}
\usage{
get_inner_neighbor(piece_id, rings)
}
\description{
Get inner ring piece ID that contains the midpoint (legacy function)
}
\keyword{internal}
