% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/piles_notation.R
\name{get_inner_pieces}
\alias{get_inner_pieces}
\title{Get inner (non-boundary) pieces (internal)}
\usage{
get_inner_pieces(puzzle_result = NULL)
}
\arguments{
\item{puzzle_result}{Puzzle result}
}
\value{
Integer vector of inner piece IDs
}
\description{
Get inner (non-boundary) pieces (internal)
}
\keyword{internal}
