% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hexagonal_neighbors.R
\name{get_inner_ring_neighbor}
\alias{get_inner_ring_neighbor}
\title{Helper function to calculate neighbors for non-center pieces}
\usage{
get_inner_ring_neighbor(piece_id, side, ring_info, rings)
}
\description{
Helper function to calculate neighbors for non-center pieces
}
\keyword{internal}
