% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/noise_fills.R
\name{get_noise_types}
\alias{get_noise_types}
\title{Get available noise types}
\usage{
get_noise_types()
}
\value{
Named character vector of noise types
}
\description{
Returns a named vector of available noise types with descriptions.
}
