% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/concentric_edge_generation.R
\name{get_outer_neighbors}
\alias{get_outer_neighbors}
\title{Get all outer ring piece IDs that share an edge with an inner ring piece}
\usage{
get_outer_neighbors(piece_id, rings)
}
\arguments{
\item{piece_id}{The inner piece ID}

\item{rings}{Total number of rings}
}
\value{
Vector of outer piece IDs that share the OUTER edge
}
\description{
Get all outer ring piece IDs that share an edge with an inner ring piece
}
\keyword{internal}
