% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/piles_notation.R
\name{get_row_pieces}
\alias{get_row_pieces}
\title{Get pieces in a specific row (rectangular puzzles) (internal)}
\usage{
get_row_pieces(row_num, puzzle_result = NULL)
}
\arguments{
\item{row_num}{Row number (1-indexed from top)}

\item{puzzle_result}{Puzzle result}
}
\value{
Integer vector of piece IDs
}
\description{
Get pieces in a specific row (rectangular puzzles) (internal)
}
\keyword{internal}
