% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hexagonal_individual_pieces.R
\name{get_segment_raw_path}
\alias{get_segment_raw_path}
\title{Get raw path from segment data}
\usage{
get_segment_raw_path(segment_data)
}
\arguments{
\item{segment_data}{Segment data structure}
}
\value{
Raw SVG path string or NULL
}
\description{
Get raw path from segment data
}
