% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/snic_puzzle.R
\name{has_magick}
\alias{has_magick}
\title{Check if magick package is available}
\usage{
has_magick()
}
\value{
TRUE if magick is installed, FALSE otherwise
}
\description{
Check if magick package is available
}
\keyword{internal}
