% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tessellation_edge_generation.R
\name{has_rcdt}
\alias{has_rcdt}
\title{Check if RCDT package is available}
\usage{
has_rcdt()
}
\value{
TRUE if RCDT is installed, FALSE otherwise
}
\description{
Check if RCDT package is available
}
\keyword{internal}
