% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/snic_puzzle.R
\name{has_snic}
\alias{has_snic}
\title{Check if snic package is available}
\usage{
has_snic()
}
\value{
TRUE if snic is installed, FALSE otherwise
}
\description{
Check if snic package is available
}
\keyword{internal}
