% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hexagonal_puzzle.R
\name{init_hex_jigsaw}
\alias{init_hex_jigsaw}
\title{Initialize hexagonal jigsaw environment}
\usage{
init_hex_jigsaw(
  seed = NULL,
  tabsize = 6,
  jitter = 5,
  diameter = 240,
  rings = 6,
  do_warp = FALSE,
  do_trunc = FALSE
)
}
\arguments{
\item{seed}{Random seed (default: random)}

\item{tabsize}{Tab size as percentage (0-100). Default: 6.}

\item{jitter}{Jitter percentage (default: 5)}

\item{diameter}{Puzzle diameter in mm (default: 240)}

\item{rings}{Number of rings (default: 6)}

\item{do_warp}{Circle warp (default: FALSE)}

\item{do_trunc}{Truncate edge pieces (default: FALSE)}
}
\description{
Initialize hexagonal jigsaw environment
}
