% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adjacency_api.R
\name{is_edge_fused}
\alias{is_edge_fused}
\title{Check if an edge is fused}
\usage{
is_edge_fused(piece_id, direction, fused_edge_data)
}
\arguments{
\item{piece_id}{Piece ID}

\item{direction}{Edge direction}

\item{fused_edge_data}{Output from compute_fused_edges()}
}
\value{
TRUE if edge is fused (internal to a meta-piece), FALSE otherwise
}
\description{
Check if an edge is fused
}
