% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jigsawR-package.R
\docType{package}
\name{jigsawR-package}
\alias{jigsawR}
\alias{jigsawR-package}
\title{jigsawR: R Translations of JavaScript Jigsaw Puzzle Generators}
\description{
R translations of Draradech's JavaScript jigsaw puzzle generators with enhanced puzzle visualization capabilities. Supports rectangular, hexagonal, concentric, Voronoi, random shape, and SNIC superpixel puzzles with high-quality SVG output, gradient backgrounds, and configurable piece separation. Direct ports of the original JavaScript algorithms maintaining mathematical precision while adding R-based visualization, tessellation, and image-aware segmentation capabilities.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/pjt222/jigsawR}
  \item Report bugs at \url{https://github.com/pjt222/jigsawR/issues}
}

}
\author{
\strong{Maintainer}: Philipp Thoss \email{ph.thoss@gmx.de} (\href{https://orcid.org/0000-0002-4672-2792}{ORCID})

}
\keyword{internal}
