% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/launch_app.R
\name{jigsawR_app}
\alias{jigsawR_app}
\title{Launch jigsawR Shiny App (alias)}
\usage{
jigsawR_app(...)
}
\arguments{
\item{...}{Arguments passed to launch_jigsaw_app()}
}
\description{
Alias for launch_jigsaw_app() for convenience
}
