% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/logging.R
\name{log_subheader}
\alias{log_subheader}
\title{Log a subsection header}
\usage{
log_subheader(text, .envir = parent.frame())
}
\arguments{
\item{text}{Subheader text}

\item{.envir}{Environment for variable lookup (defaults to parent frame)}
}
\description{
Log a subsection header
}
\keyword{internal}
