% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adjacency_api.R
\name{merge_fusion_groups}
\alias{merge_fusion_groups}
\title{Merge overlapping fusion groups}
\usage{
merge_fusion_groups(fusion_groups)
}
\arguments{
\item{fusion_groups}{List of integer vectors representing fusion groups}
}
\value{
List of merged integer vectors (sorted, deduplicated)
}
\description{
Uses Union-Find algorithm to merge fusion groups that share common pieces.
For example, (7,8) and (5,8) share piece 8, so they merge into (5,7,8).
}
\examples{
\dontrun{
merge_fusion_groups(list(c(7,8), c(5,8)))    # -> list(c(5,7,8))
merge_fusion_groups(list(c(1,2), c(2,3)))    # -> list(c(1,2,3))
merge_fusion_groups(list(c(1,2), c(3,4)))    # -> list(c(1,2), c(3,4))
}

}
