% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/noise_fills.R
\name{noise_to_colors}
\alias{noise_to_colors}
\title{Convert noise values to color matrix}
\usage{
noise_to_colors(
  noise_matrix,
  width,
  height,
  color_low,
  color_high,
  color_mid = NULL
)
}
\arguments{
\item{noise_matrix}{Matrix of normalized noise values (height x width)}

\item{width}{Image width}

\item{height}{Image height}

\item{color_low}{Color for value 0}

\item{color_high}{Color for value 1}

\item{color_mid}{Optional color for value 0.5}
}
\value{
3D array (height x width x 3) of RGB values [0,1]
}
\description{
Maps normalized noise values [0,1] to colors using a gradient.
}
\keyword{internal}
