% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/piles_notation.R
\name{parse_fusion}
\alias{parse_fusion}
\title{Enhanced fusion input parser with PILES support}
\usage{
parse_fusion(input, puzzle_result = NULL, auto_merge = TRUE)
}
\arguments{
\item{input}{Fusion input in any format:
\itemize{
  \item PILES string: "1-2-3,4-5"
  \item Legacy string: "(1,2),(3,4)"
  \item List of vectors: list(c(1,2), c(3,4))
}}

\item{puzzle_result}{Optional puzzle result for validation}

\item{auto_merge}{If TRUE, merge overlapping groups}
}
\value{
List of integer vectors (fusion groups)
}
\description{
Extended version of parse_fusion_input that also accepts PILES notation.
Automatically detects format based on input.
}
