% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bezier_utils.R
\name{parse_svg_path}
\alias{parse_svg_path}
\title{Parse SVG path string into structured segments}
\usage{
parse_svg_path(path_string)
}
\arguments{
\item{path_string}{SVG path d attribute string}
}
\value{
List of path segments
}
\description{
Parse SVG path string into structured segments
}
