% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hexagonal_piece_extraction.R
\name{points_match}
\alias{points_match}
\title{Check if two points match within tolerance}
\usage{
points_match(p1, p2, tolerance = 0.1)
}
\arguments{
\item{p1}{Point 1 c(x, y)}

\item{p2}{Point 2 c(x, y)}

\item{tolerance}{Matching tolerance}
}
\value{
TRUE if points match
}
\description{
Check if two points match within tolerance
}
