% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hexagonal_puzzle.R
\name{print_hex_puzzle_info}
\alias{print_hex_puzzle_info}
\title{Print hexagonal puzzle parameters}
\usage{
print_hex_puzzle_info(puzzle_data)
}
\arguments{
\item{puzzle_data}{Output from generate_hex_jigsaw_svg()}
}
\description{
Print hexagonal puzzle parameters
}
