% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rectangular_puzzle.R
\name{print_puzzle_info}
\alias{print_puzzle_info}
\title{Print puzzle parameters}
\usage{
print_puzzle_info(puzzle_data)
}
\arguments{
\item{puzzle_data}{Output from generate_jigsaw_svg()}
}
\description{
Print puzzle parameters
}
