% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hexagonal_individual_pieces.R
\name{process_segment_for_piece}
\alias{process_segment_for_piece}
\title{Process segment for piece-specific use}
\usage{
process_segment_for_piece(raw_segment, segment_info, piece_pos, segment_index)
}
\arguments{
\item{raw_segment}{Raw SVG path segment}

\item{segment_info}{Segment metadata}

\item{piece_pos}{Piece position data}

\item{segment_index}{Index of this segment in the boundary}
}
\value{
Processed SVG path segment
}
\description{
Process segment for piece-specific use
}
