% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/snic_puzzle.R
\name{process_snic_fusion}
\alias{process_snic_fusion}
\title{Process fusion groups for SNIC puzzle}
\usage{
process_snic_fusion(puzzle_result, fusion_groups, fusion_style, fusion_opacity)
}
\arguments{
\item{puzzle_result}{Puzzle result}

\item{fusion_groups}{Fusion specification}

\item{fusion_style}{Style for fused edges}

\item{fusion_opacity}{Opacity for fusion regions}
}
\value{
Updated fusion data or NULL
}
\description{
Process fusion groups for SNIC puzzle
}
\keyword{internal}
