% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/voronoi_puzzle.R
\name{process_voronoi_fusion}
\alias{process_voronoi_fusion}
\title{Process fusion groups for Voronoi puzzle}
\usage{
process_voronoi_fusion(
  puzzle_result,
  fusion_groups,
  fusion_style,
  fusion_opacity
)
}
\arguments{
\item{puzzle_result}{Puzzle result}

\item{fusion_groups}{Fusion specification}

\item{fusion_style}{Style for fused edges}

\item{fusion_opacity}{Opacity for fusion regions}
}
\value{
Updated fusion data
}
\description{
Process fusion groups for Voronoi puzzle
}
\keyword{internal}
