% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hexagonal_individual_pieces.R
\name{rebuild_segment_path}
\alias{rebuild_segment_path}
\title{Rebuild SVG path segment from parsed data}
\usage{
rebuild_segment_path(segment_data)
}
\arguments{
\item{segment_data}{Single parsed segment}
}
\value{
SVG path string
}
\description{
Rebuild SVG path segment from parsed data
}
