% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adjacency_api.R
\name{rect_pos_to_index}
\alias{rect_pos_to_index}
\title{Convert rectangular grid position to piece index}
\usage{
rect_pos_to_index(xi, yi, xn)
}
\arguments{
\item{xi}{Column (0-based)}

\item{yi}{Row (0-based)}

\item{xn}{Number of columns}
}
\value{
1-based piece index
}
\description{
Convert rectangular grid position to piece index
}
\keyword{internal}
