% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/individual_pieces_correct.R
\name{reverse_bezier}
\alias{reverse_bezier}
\title{Reverse a bezier curve segment}
\usage{
reverse_bezier(start_point, curve_data)
}
\description{
Reverses the direction of a bezier curve for complementary edges
}
\keyword{internal}
