% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bezier_utils.R
\name{reverse_path_segments}
\alias{reverse_path_segments}
\title{Reverse a series of bezier curve segments}
\usage{
reverse_path_segments(segments, start_point)
}
\arguments{
\item{segments}{List of parsed path segments to reverse}

\item{start_point}{Starting point for the reversed path}
}
\value{
Reversed SVG path string
}
\description{
Reverse a series of bezier curve segments
}
