% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/individual_pieces.R
\name{save_combined_pieces_svg}
\alias{save_combined_pieces_svg}
\title{Save combined view of all pieces}
\usage{
save_combined_pieces_svg(pieces, width, height, output_dir, palette = NULL)
}
\description{
Save combined view of all pieces
}
\keyword{internal}
