% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/svg_utils.R
\name{save_enhanced_svg}
\alias{save_enhanced_svg}
\title{Save SVG content to file}
\usage{
save_enhanced_svg(svg_content, svg_file)
}
\arguments{
\item{svg_content}{SVG content string}

\item{svg_file}{Output filename for SVG}
}
\description{
Save SVG content to file
}
