% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gradient_background.R
\name{save_gradient_background}
\alias{save_gradient_background}
\title{Save gradient background to file}
\usage{
save_gradient_background(gradient_plot, background_file, size_px = 2000)
}
\arguments{
\item{gradient_plot}{ggplot2 object from create_gradient_circle_png()}

\item{background_file}{Output filename for background}

\item{size_px}{Size in pixels}
}
\description{
Save gradient background to file
}
