% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hexagonal_puzzle.R
\name{save_hex_jigsaw_svg}
\alias{save_hex_jigsaw_svg}
\title{Save hexagonal SVG to file}
\usage{
save_hex_jigsaw_svg(puzzle_data, filename = "hex_jigsaw.svg")
}
\arguments{
\item{puzzle_data}{Output from generate_hex_jigsaw_svg()}

\item{filename}{Output filename (default: "hex_jigsaw.svg")}
}
\description{
Save hexagonal SVG to file
}
