% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hexagonal_individual_pieces.R
\name{save_hexagonal_individual_pieces}
\alias{save_hexagonal_individual_pieces}
\title{Save individual hexagonal pieces as separate files}
\usage{
save_hexagonal_individual_pieces(pieces, puzzle_struct, output_dir)
}
\arguments{
\item{pieces}{List of individual pieces}

\item{puzzle_struct}{Puzzle structure}

\item{output_dir}{Output directory}
}
\value{
Vector of saved filenames
}
\description{
Save individual hexagonal pieces as separate files
}
