% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/individual_pieces.R
\name{save_individual_piece_svg}
\alias{save_individual_piece_svg}
\title{Save individual piece as SVG file}
\usage{
save_individual_piece_svg(xi, yi, piece_path, width, height, output_dir)
}
\description{
Save individual piece as SVG file
}
\keyword{internal}
