% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rectangular_puzzle.R
\name{save_jigsaw_svg}
\alias{save_jigsaw_svg}
\title{Save SVG to file}
\usage{
save_jigsaw_svg(puzzle_data, filename = "jigsaw.svg")
}
\arguments{
\item{puzzle_data}{Output from generate_jigsaw_svg()}

\item{filename}{Output filename (default: "jigsaw.svg")}
}
\description{
Save SVG to file
}
