% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hexagonal_piece_extraction.R
\name{split_path_by_move}
\alias{split_path_by_move}
\title{Split SVG path by M commands to get individual segments}
\usage{
split_path_by_move(svg_path)
}
\arguments{
\item{svg_path}{Complete SVG path string}
}
\value{
List of individual path segments with their data
}
\description{
Split SVG path by M commands to get individual segments
}
