% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/piles_notation.R
\name{to_piles}
\alias{to_piles}
\title{Convert fusion groups to PILES notation}
\usage{
to_piles(fusion_groups, puzzle_result = NULL, compact = TRUE)
}
\arguments{
\item{fusion_groups}{List of integer vectors}

\item{puzzle_result}{Optional puzzle result for structure-aware serialization}

\item{compact}{If TRUE, use range notation where possible}
}
\value{
Character string in PILES notation
}
\description{
Serializes a list of fusion groups back to PILES string format.
}
\examples{
to_piles(list(c(1, 2), c(3, 4, 5)))
# Returns: "1-2,3-4-5"

to_piles(list(1:6))
# Returns: "1:6" (with compact=TRUE)

}
