% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/piles_notation.R
\name{tokenize_piles}
\alias{tokenize_piles}
\title{Tokenize PILES string (internal)}
\usage{
tokenize_piles(str)
}
\arguments{
\item{str}{PILES string}
}
\value{
Character vector of tokens
}
\description{
Tokenize PILES string (internal)
}
\keyword{internal}
