% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/individual_pieces_correct.R
\name{transform_for_complement}
\alias{transform_for_complement}
\title{Transform coordinates for complementary edges}
\usage{
transform_for_complement(coords, axis = "x", center = 100)
}
\description{
Flips coordinates around a centerline for complementary puzzle edges
}
\keyword{internal}
