% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/piece_positioning.R
\name{translate_piece}
\alias{translate_piece}
\title{Translate a piece by dx, dy}
\usage{
translate_piece(piece, dx, dy)
}
\arguments{
\item{piece}{Piece object}

\item{dx}{X translation}

\item{dy}{Y translation}
}
\value{
Translated piece
}
\description{
Translate a piece by dx, dy
}
\keyword{internal}
