% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/piece_positioning.R
\name{translate_segments}
\alias{translate_segments}
\title{Translate parsed segments by dx, dy}
\usage{
translate_segments(segments, dx, dy)
}
\arguments{
\item{segments}{List of parsed SVG segments}

\item{dx}{X translation}

\item{dy}{Y translation}
}
\value{
Translated segments
}
\description{
Translate parsed segments by dx, dy
}
\keyword{internal}
