% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/voronoi_puzzle.R
\name{translate_voronoi_path}
\alias{translate_voronoi_path}
\title{Translate Voronoi path by offset}
\usage{
translate_voronoi_path(path, dx, dy)
}
\arguments{
\item{path}{SVG path string}

\item{dx}{X offset}

\item{dy}{Y offset}
}
\value{
Translated path string
}
\description{
Uses the existing translate_svg_path function from piece_positioning.R
}
\keyword{internal}
