% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hexagonal_individual_pieces.R
\name{create_hexagonal_individual_pieces_svg}
\alias{create_hexagonal_individual_pieces_svg}
\title{Create combined SVG for all individual pieces}
\usage{
create_hexagonal_individual_pieces_svg(
  pieces,
  puzzle_struct,
  colors,
  stroke_width
)
}
\arguments{
\item{pieces}{List of individual pieces}

\item{puzzle_struct}{Original puzzle structure}

\item{colors}{Color palette}

\item{stroke_width}{Line width}
}
\value{
Complete SVG string
}
\description{
Create combined SVG for all individual pieces
}
