% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/individual_pieces_correct.R
\name{generate_individual_pieces}
\alias{generate_individual_pieces}
\title{Generate individual pieces from a full puzzle}
\usage{
generate_individual_pieces(
  seed = 42,
  xn = 2,
  yn = 2,
  width = 200,
  height = 200,
  output_dir = "output"
)
}
\arguments{
\item{seed}{Random seed}

\item{xn}{Number of columns}

\item{yn}{Number of rows}

\item{width}{Puzzle width}

\item{height}{Puzzle height}

\item{output_dir}{Output directory}
}
\value{
List of piece data
}
\description{
This is the general function that will work for any size puzzle.
Currently implements the algorithm for 2x2, can be extended.
}
