% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main_generator.R
\name{generate_puzzle_variations}
\alias{generate_puzzle_variations}
\title{Generate multiple puzzle variations}
\usage{
generate_puzzle_variations(variations)
}
\arguments{
\item{variations}{List of variation parameters, each containing seed, rings, etc.}
}
\value{
List of results for each variation
}
\description{
Generate multiple puzzle variations
}
\examples{
\donttest{
variations <- list(
  list(seed = 42, rings = 3, base_filename = "puzzle_v1"),
  list(seed = 99, rings = 4, base_filename = "puzzle_v2")
)
results <- generate_puzzle_variations(variations)
}

}
