% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geom_puzzle.R
\docType{data}
\name{GeomPuzzle}
\alias{GeomPuzzle}
\title{Puzzle piece geom}
\format{
An object of class \code{GeomPuzzle} (inherits from \code{Geom}, \code{ggproto}, \code{gg}) of length 5.
}
\usage{
GeomPuzzle
}
\description{
Renders puzzle pieces as filled polygons. Works with StatPuzzle to
generate piece geometry from data.
}
\keyword{datasets}
