% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adjacency_api.R
\name{get_piece_fusion_group}
\alias{get_piece_fusion_group}
\title{Get fusion group for a piece}
\usage{
get_piece_fusion_group(piece_id, fused_edge_data)
}
\arguments{
\item{piece_id}{Piece ID}

\item{fused_edge_data}{Output from compute_fused_edges()}
}
\value{
Fusion group index (integer), or NA if piece is not in any group
}
\description{
Get fusion group for a piece
}
