% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/config_utils.R
\name{get_puzzle_colors}
\alias{get_puzzle_colors}
\title{Get viridis color palette}
\usage{
get_puzzle_colors(n, palette = NULL, invert = FALSE)
}
\arguments{
\item{n}{Number of colors to generate}

\item{palette}{Palette name (black, viridis, magma, plasma, inferno, cividis, mako, rocket, turbo)}

\item{invert}{Logical, if TRUE reverses the palette direction (default: FALSE)}
}
\value{
Character vector of hex colors
}
\description{
Returns colors from specified viridis palette
}
