% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rectangular_puzzle.R
\name{init_jigsaw}
\alias{init_jigsaw}
\title{Initialize jigsaw environment with global variables}
\usage{
init_jigsaw(
  seed = NULL,
  tabsize = 20,
  jitter = 4,
  width = 300,
  height = 200,
  radius = 2,
  xn = 15,
  yn = 10
)
}
\arguments{
\item{seed}{Random seed (default: random)}

\item{tabsize}{Tab size percentage (default: 20)}

\item{jitter}{Jitter percentage (default: 4)}

\item{width}{Puzzle width in mm (default: 300)}

\item{height}{Puzzle height in mm (default: 200)}

\item{radius}{Corner radius in mm (default: 2.0)}

\item{xn}{Number of columns (default: 15)}

\item{yn}{Number of rows (default: 10)}
}
\description{
Initialize jigsaw environment with global variables
}
