% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/noise_fills.R
\name{is_noise_fill_spec}
\alias{is_noise_fill_spec}
\title{Check if object is a noise fill specification}
\usage{
is_noise_fill_spec(x)
}
\arguments{
\item{x}{Object to check}
}
\value{
TRUE if x is a noise fill spec, FALSE otherwise
}
\description{
Check if object is a noise fill specification
}
\examples{
is_noise_fill_spec(list(type = "noise", noise_type = "perlin"))
is_noise_fill_spec("#FF0000")
is_noise_fill_spec(list(type = "gradient"))

}
