% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/puzzle_separation.R
\name{translate_svg_path}
\alias{translate_svg_path}
\title{Apply offset to SVG path coordinates}
\usage{
translate_svg_path(path_string, x_offset, y_offset)
}
\arguments{
\item{path_string}{SVG path d attribute string}

\item{x_offset}{X translation amount}

\item{y_offset}{Y translation amount}
}
\value{
Translated SVG path string
}
\description{
Translates all coordinates in an SVG path by the specified offset.
}
