% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/piles_notation.R
\name{validate_piles_syntax}
\alias{validate_piles_syntax}
\title{Validate PILES notation syntax}
\usage{
validate_piles_syntax(piles)
}
\arguments{
\item{piles}{Character string in PILES notation}
}
\value{
List with:
  \describe{
    \item{valid}{TRUE if syntax is valid}
    \item{message}{Error message if invalid}
    \item{warnings}{Vector of warning messages}
  }
}
\description{
Checks if a PILES string has valid syntax without evaluating special keywords.
}
\examples{
validate_piles_syntax("1-2-3,4-5")
validate_piles_syntax("1:6")
validate_piles_syntax("1-2(-3)-4")
validate_piles_syntax("((unbalanced")

}
